requirements.txt
openai
langchain
pandas
langchain-openai
python-dotenv
langchain-community
ollama